# ⚡️ improve perf of ws client