--- a/SurrealDb.Net/SurrealDb.Net.csproj
+++ b/SurrealDb.Net/SurrealDb.Net.csproj
@@ -11,7 +11,7 @@
 
   <ItemGroup>
 	<PackageReference Include="Microsoft.Extensions.Http" Version="8.0.0" />
-	<PackageReference Include="Microsoft.IO.RecyclableMemoryStream" Version="2.3.2" />
+	<PackageReference Include="Microsoft.IO.RecyclableMemoryStream" Version="3.0.0" />
 	<PackageReference Include="Microsoft.Spatial" Version="7.18.0" />
 	<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
 	<PackageReference Include="Semver" Version="2.3.0" />
@@ -19,7 +19,7 @@
 	<PackageReference Include="System.Collections.Immutable" Version="8.0.0" />
 	<PackageReference Include="System.Text.Json" Version="8.0.0" />
 	<PackageReference Include="SystemTextJsonPatch" Version="3.0.1" />
-	<PackageReference Include="Websocket.Client" Version="5.0.0" />
+	<PackageReference Include="Websocket.Client" Version="5.1.1" />
   </ItemGroup>
 
   <ItemGroup>