Performance improvement on the WS client implementation following the PR https://github.com/Marfusios/websocket-client/pull/133 [Waiting for merge]

Here is the output benchmark after optimization:

### Cold Start
| Method                | Mean        | Error     | StdDev    | Allocated |
|---------------------- |------------:|----------:|----------:|----------:|
| HttpConstructor       |    15.13 ms |  0.052 ms |  0.046 ms |  11.61 KB |
| HttpStaticConstructor |    15.01 ms |  0.102 ms |  0.091 ms |  11.87 KB |
| WsConstructor         | 2,050.25 ms | 12.877 ms | 12.045 ms | 224.45 KB |
| WsStaticConstructor   | 2,045.68 ms | 14.527 ms | 13.588 ms | 226.04 KB |

### Select
| Method                | Mean        | Error     | StdDev   | Allocated |
|---------------------- |------------:|----------:|---------:|----------:|
| Http                  | 14,952.3 μs | 105.25 μs | 98.45 μs |   12.2 KB |
| HttpWithClientFactory | 14,910.2 μs |  67.69 μs | 60.01 μs |  13.71 KB |
| WsText                |    630.4 μs |   9.40 μs |  8.79 μs |   9.99 KB |
| WsBinary              |          NA |        NA |       NA |        NA |

### Create
| Method                | Mean        | Error    | StdDev   | Gen0   | Allocated |
|---------------------- |------------:|---------:|---------:|-------:|----------:|
| Http                  | 15,199.6 μs | 82.90 μs | 73.49 μs |      - |  57.11 KB |
| HttpWithClientFactory | 15,193.0 μs | 60.43 μs | 56.53 μs |      - |  53.94 KB |
| WsText                |    754.3 μs |  4.76 μs |  3.97 μs | 2.9297 |  48.13 KB |
| WsBinary              |          NA |       NA |       NA |     NA |        NA |

### Upsert
| Method                | Mean        | Error     | StdDev    | Gen0   | Allocated |
|---------------------- |------------:|----------:|----------:|-------:|----------:|
| Http                  | 15,082.5 μs |  92.14 μs |  76.94 μs |      - |  27.55 KB |
| HttpWithClientFactory | 15,103.9 μs | 129.15 μs | 107.84 μs |      - |  27.85 KB |
| WsText                |    711.3 μs |   6.20 μs |   5.80 μs | 0.9766 |  20.04 KB |
| WsBinary              |          NA |        NA |        NA |     NA |        NA |

### Delete
| Method                | Mean        | Error     | StdDev    | Allocated |
|---------------------- |------------:|----------:|----------:|----------:|
| Http                  | 14,914.6 μs | 111.77 μs | 104.55 μs |   8.37 KB |
| HttpWithClientFactory | 14,888.5 μs |  52.98 μs |  46.97 μs |  10.11 KB |
| WsText                |    574.8 μs |   7.29 μs |   5.69 μs |   3.91 KB |
| WsBinary              |          NA |        NA |        NA |        NA |

### Query
| Method                | Mean      | Error     | StdDev    | Allocated |
|---------------------- |----------:|----------:|----------:|----------:|
| Http                  | 15.721 ms | 0.2373 ms | 0.2220 ms |  17.33 KB |
| HttpWithClientFactory | 15.627 ms | 0.2036 ms | 0.1905 ms |  17.78 KB |
| WsText                |  1.292 ms | 0.0107 ms | 0.0100 ms |  15.58 KB |
| WsBinary              |        NA |        NA |        NA |        NA |

### Scenario
| Method                | Mean      | Error    | StdDev   | Allocated |
|---------------------- |----------:|---------:|---------:|----------:|
| Http                  | 192.02 ms | 1.504 ms | 1.333 ms | 173.23 KB |
| HttpWithClientFactory | 192.38 ms | 2.491 ms | 2.330 ms | 199.02 KB |
| WsText                |  10.85 ms | 0.105 ms | 0.098 ms |  99.29 KB |
| WsBinary              |        NA |       NA |       NA |        NA |

Note 1: Memory allocation reduced up to 50% on WS method calls

Note 2: WS method calls slightly faster (due to less memory allocation I suppose), around 1~3% faster than previous version

Note 3: WS "cold start" requires a significant amount of memory, due to the use of `RecyclableMemoryStreamManager` on the underlying websocket-client. Not surprising, as a WS client should make multiple calls in general. Also note that the `RecyclableMemoryStreamManager` is static, so it doesn't matter how many WS clients you use.