@Odonno 2024-02-29
Updated benchmarks:


### Cold Start
| Method          | Mean     | Error    | StdDev   | Allocated |
|---------------- |---------:|---------:|---------:|----------:|
| HttpConstructor | 14.58 ms | 0.272 ms | 0.267 ms |   8.68 KB |
| WsConstructor   | 15.81 ms | 0.279 ms | 0.261 ms | 185.08 KB |

### Select
| Method                | Mean      | Error     | StdDev    | Gen0    | Gen1    | Allocated |
|---------------------- |----------:|----------:|----------:|--------:|--------:|----------:|
| Http                  | 15.672 ms | 0.0517 ms | 0.0432 ms |       - |       - | 271.45 KB |
| HttpWithClientFactory | 15.695 ms | 0.0866 ms | 0.0768 ms |       - |       - |  267.5 KB |
| WsText                |  7.110 ms | 0.0933 ms | 0.0827 ms | 62.5000 | 31.2500 | 1993.8 KB |
| WsBinary              |          NA |        NA |       NA |        NA |

### Create
| Method                | Mean        | Error     | StdDev    | Gen0   | Allocated |
|---------------------- |------------:|----------:|----------:|-------:|----------:|
| Http                  | 14,798.4 us | 141.09 us | 117.81 us |      - |  51.82 KB |
| HttpWithClientFactory | 14,967.4 us | 270.75 us | 240.02 us |      - |   52.4 KB |
| WsText                |    289.2 us |   4.39 us |   3.90 us | 2.4414 |  45.89 KB |
| WsBinary              |          NA |       NA |       NA |     NA |        NA |

### Upsert
| Method                | Mean        | Error     | StdDev    | Gen0   | Allocated |
|---------------------- |------------:|----------:|----------:|-------:|----------:|
| Http                  | 15,043.8 us | 122.41 us | 114.50 us |      - |   23.5 KB |
| HttpWithClientFactory | 15,162.6 us | 273.59 us | 242.53 us |      - |  25.43 KB |
| WsText                |    251.8 us |   3.60 us |   3.37 us | 0.9766 |  18.64 KB |
| WsBinary              |          NA |        NA |        NA |     NA |        NA |

### Delete
| Method                | Mean        | Error     | StdDev    | Allocated |
|---------------------- |------------:|----------:|----------:|----------:|
| Http                  | 14,469.3 us | 118.94 us | 105.44 us |   6.15 KB |
| HttpWithClientFactory | 14,385.1 us | 107.84 us | 100.88 us |   7.34 KB |
| WsText                |    150.9 us |   2.17 us |   2.03 us |   3.51 KB |
| WsBinary              |          NA |        NA |        NA |        NA |

### Query
| Method                | Mean      | Error     | StdDev    | Gen0    | Gen1    | Allocated |
|---------------------- |----------:|----------:|----------:|--------:|--------:|----------:|
| Http                  | 22.681 ms | 0.1913 ms | 0.1790 ms | 62.5000 | 31.2500 |   2.57 MB |
| HttpWithClientFactory | 22.513 ms | 0.0889 ms | 0.0742 ms | 62.5000 | 31.2500 |   2.64 MB |
| WsText                |  8.750 ms | 0.0479 ms | 0.0448 ms | 62.5000 | 31.2500 |   2.63 MB |
| WsBinary              |        NA |        NA |        NA |        NA |

### Scenario
| Method                | Mean       | Error     | StdDev    | Gen0   | Allocated |
|---------------------- |-----------:|----------:|----------:|-------:|----------:|
| Http                  | 181.609 ms | 2.3738 ms | 2.1043 ms |      - | 134.09 KB |
| HttpWithClientFactory | 182.404 ms | 0.8664 ms | 0.7235 ms |      - | 151.79 KB |
| WsText                |   3.080 ms | 0.0592 ms | 0.0634 ms | 3.9063 |  85.54 KB |
| WsBinary              |        NA |       NA |       NA |        NA |